fix(plugin): wp_filesystem undefined fix
On my WordPress Multisite installation, I encountered an issue where the WP_Filesystem function was undefined. This error was causing all of my posts and custom post types not to update or save because the function wasn't included. Fixing the problem required checking if the function was defined and if not, then including the appropriate core file.